rearrange generatePartnerNetworkInviteEmailAction
